pyproject.toml
[project]
name = "mul-agentic-chatbot-with-langgraph"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.8",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.8",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "slowapi>=0.1.9",
    "tavily-python>=0.7.21",
    "uvicorn>=0.40.0",
]
